@@ -3,7 +3,7 @@
     "description": "PHP refactoring and intellisense tool for text editors",
     "license": "MIT",
     "require": {
-        "php": "^8.0",
+        "php": "^8.1",
         "ext-pcntl": "*",
         "ext-posix": "*",
         "composer/package-versions-deprecated": "^1.0",
@@ -35,7 +35,7 @@
     },
     "require-dev": {
         "dantleech/what-changed": "~0.4",
-        "friendsofphp/php-cs-fixer": "^3.15",
+        "friendsofphp/php-cs-fixer": "^3.32",
         "phpstan/phpstan": "^1.0",
         "phpunit/phpunit": "^9.0",
         "phpspec/prophecy-phpunit": "^2.0",
@@ -86,7 +86,7 @@
     },
     "config": {
         "platform": {
-            "php": "8.0.0"
+            "php": "8.1.0"
         },
         "allow-plugins": {
             "dantleech/what-changed": true,
